Extract view helper functions from templates to improve separation of concerns
