+++ app/Controllers/BaseController.php
@@ -39,7 +39,7 @@ abstract class BaseController extends Controller
      *
      * @var list<string>
      */
-    protected $helpers = [];
+    protected $helpers = ['view'];
 
     /**
      * Constructor.
